Use frequencies as input data
The analyzer is dependant on the samplerate of the audio context, so we
calculate the max frequency as half that of the samplerate.

Data lookups now take frequency range into consideration. Old float
values were based on the assumption of 48khz audio, so multiplying old
input values by 48000 will give you the frequency you desire.