Add band bleed parameter to createBands()
bleed allows for the band size to enlarge (both start and end) past the
normal constraints when creating a range of bands. The default behaviour
is 0.

A value of 0.5 would have each band overlapping with the previous and
next by half the size of the band.

This should help use-cases where you want a smoother path between band
values.